Fix type in shader

+++ data/graphic/shaders/scalable_kinect_fusion/rasterization_reverse/cs_gather.glsl
@@ -15,7 +15,7 @@ layout(binding = 1, r8ui) uniform uimage3D cs_Volume;
 layout (local_size_x = 16, local_size_y = 16, local_size_z = 1) in;
 void main()
 {
-    imageStore(cs_Volume, ivec3(gl_GlobalInvocationID), vec4(0.0f));
+    imageStore(cs_Volume, ivec3(gl_GlobalInvocationID), uvec4(0));
 }
 
 #endif // __INCLUDE_CS_GATHER_GATHER_GLSL__
